+++ lib/teamtailor/parser/reject_reason.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require 'teamtailor/record'
+
+module Teamtailor
+  class RejectReason < Record
+  end
+end